+++ libos/src/libos_init.c
@@ -371,6 +371,8 @@ static int read_environs(const char* const* envp) {
     } while (0)
 
 noreturn void libos_init(const char* const* argv, const char* const* envp) {
+    int ret;
+
     g_pal_public_state = PalGetPalPublicState();
     assert(g_pal_public_state);
 
@@ -410,12 +412,18 @@ noreturn void libos_init(const char* const* argv, const char* const* envp) {
     RUN_INIT(init_dcache);
     RUN_INIT(init_handle);
 
+    ret = print_warnings_on_insecure_configs(!!g_pal_public_state->parent_process);
+    if (ret < 0) {
+        log_error("Cannot parse the manifest (while checking for insecure configurations)");
+        PalProcessExit(1);
+    }
+
     log_debug("LibOS loaded at %p, ready to initialize", &__load_address);
 
     if (g_pal_public_state->parent_process) {
         struct checkpoint_hdr hdr;
 
-        int ret = read_exact(g_pal_public_state->parent_process, &hdr, sizeof(hdr));
+        ret = read_exact(g_pal_public_state->parent_process, &hdr, sizeof(hdr));
         if (ret < 0) {
             log_error("libos_init: failed to read the whole checkpoint header: %s",
                       unix_strerror(ret));
@@ -469,7 +477,7 @@ noreturn void libos_init(const char* const* argv, const char* const* envp) {
     RUN_INIT(init_ipc_worker);
 
     if (g_pal_public_state->parent_process) {
-        int ret = connect_to_process(g_process_ipc_ids.parent_vmid);
+        ret = connect_to_process(g_process_ipc_ids.parent_vmid);
         if (ret < 0) {
             log_error("libos_init: failed to establish IPC connection to parent: %s",
                       unix_strerror(ret));
